OBSDATA-483: Address Xavier's comments
