+++ extensions-contrib/opencensus-extensions/src/main/java/org/apache/druid/data/input/opencensus/protobuf/OpenCensusProtobufInputRowParser.java
@@ -110,7 +110,6 @@ public class OpenCensusProtobufInputRowParser implements ByteBufferInputRowParse
         parseSpec.getDimensionsSpec(),
         settableByteEntity,
         metricDimension,
-        null,
         metricLabelPrefix,
         resourceLabelPrefix
     ).readAsList();
