+++ extensions-contrib/opencensus-extensions/src/main/java/org/apache/druid/data/input/opencensus/protobuf/OpenCensusProtobufReader.java
@@ -19,6 +19,8 @@
 
 package org.apache.druid.data.input.opencensus.protobuf;
 
+import com.google.common.base.Supplier;
+import com.google.common.base.Suppliers;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
@@ -30,11 +32,9 @@ import io.opencensus.proto.metrics.v1.TimeSeries;
 import org.apache.druid.data.input.InputEntityReader;
 import org.apache.druid.data.input.InputRow;
 import org.apache.druid.data.input.InputRowListPlusRawValues;
-import org.apache.druid.data.input.KafkaUtils;
 import org.apache.druid.data.input.MapBasedInputRow;
 import org.apache.druid.data.input.impl.ByteEntity;
 import org.apache.druid.data.input.impl.DimensionsSpec;
-import org.apache.druid.data.input.opentelemetry.protobuf.OpenTelemetryMetricsProtobufReader;
 import org.apache.druid.indexing.seekablestream.SettableByteEntity;
 import org.apache.druid.java.util.common.CloseableIterators;
 import org.apache.druid.java.util.common.parsers.CloseableIterator;
@@ -42,43 +42,29 @@ import org.apache.druid.java.util.common.parsers.ParseException;
 import org.apache.druid.utils.CollectionUtils;
 
 import java.io.IOException;
-import java.lang.invoke.MethodHandle;
-import java.nio.ByteBuffer;
-import java.nio.ByteOrder;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-
-import static org.apache.druid.data.input.opencensus.protobuf.OpenCensusProtobufReader.ProtobufReader.OPEN_CENSUS;
-import static org.apache.druid.data.input.opencensus.protobuf.OpenCensusProtobufReader.ProtobufReader.OPEN_TELEMETRY;
 
 public class OpenCensusProtobufReader implements InputEntityReader
 {
   private static final String SEPARATOR = "-";
   private static final String VALUE_COLUMN = "value";
-  private static final String VERSION_HEADER_KEY = "v";
-  private static final int OPENTELEMETRY_FORMAT_VERSION = 1;
 
   private final DimensionsSpec dimensionsSpec;
   private final SettableByteEntity<? extends ByteEntity> source;
   private final String metricDimension;
-  private final String valueDimension;
   private final String metricLabelPrefix;
   private final String resourceLabelPrefix;
-
-  enum ProtobufReader {
-    OPEN_CENSUS,
-    OPEN_TELEMETRY
-  }
 
   public OpenCensusProtobufReader(
       DimensionsSpec dimensionsSpec,
       SettableByteEntity<? extends ByteEntity> source,
       String metricDimension,
-      String valueDimension,
       String metricLabelPrefix,
       String resourceLabelPrefix
   )
@@ -86,7 +72,6 @@ public class OpenCensusProtobufReader implements InputEntityReader
     this.dimensionsSpec = dimensionsSpec;
     this.source = source;
     this.metricDimension = metricDimension;
-    this.valueDimension = valueDimension;
     this.metricLabelPrefix = metricLabelPrefix;
     this.resourceLabelPrefix = resourceLabelPrefix;
   }
@@ -95,69 +80,22 @@ public class OpenCensusProtobufReader implements InputEntityReader
   {
     void addRow(long millis, String metricName, Object value);
   }
-
   @Override
-  public CloseableIterator<InputRow> read() throws IOException
+  public CloseableIterator<InputRow> read()
   {
-    return newReader(whichReader()).read();
-  }
-
-  class OpenCensusProtobufReaderInternal implements InputEntityReader
-  {
-    @Override
-    public CloseableIterator<InputRow> read()
-    {
-      return CloseableIterators.withEmptyBaggage(readAsList().iterator());
-    }
-
-    @Override
-    public CloseableIterator<InputRowListPlusRawValues> sample() throws IOException
-    {
-      return OpenCensusProtobufReader.this.sample();
-    }
-  }
-
-  public InputEntityReader newReader(ProtobufReader which)
-  {
-    switch(which) {
-      case OPEN_TELEMETRY:
-        return new OpenTelemetryMetricsProtobufReader(
-            dimensionsSpec,
-            source,
-            metricDimension,
-            valueDimension,
-            metricLabelPrefix,
-            resourceLabelPrefix
-        );
-      case OPEN_CENSUS:
-      default:
-        return new OpenCensusProtobufReaderInternal();
-    }
-  }
-
-  public ProtobufReader whichReader()
-  {
-    // assume InputEntity is always defined in a single classloader (the kafka-indexing-service classloader)
-    // so we only have to look it up once. To be completely correct we should cache the method based on classloader
-    MethodHandle getHeaderMethod = KafkaUtils.lookupGetHeaderMethod(
-        source.getEntity().getClass().getClassLoader(),
-        VERSION_HEADER_KEY
-    );
-
-    try {
-      byte[] versionHeader = (byte[]) getHeaderMethod.invoke(source.getEntity());
-      if (versionHeader != null) {
-        int version =
-            ByteBuffer.wrap(versionHeader).order(ByteOrder.LITTLE_ENDIAN).getInt();
-        if (version == OPENTELEMETRY_FORMAT_VERSION) {
-          return OPEN_TELEMETRY;
-        }
+    Supplier<Iterator<InputRow>> supplier = Suppliers.memoize(() -> readAsList().iterator());
+    return CloseableIterators.withEmptyBaggage(new Iterator<InputRow>() {
+      @Override
+      public boolean hasNext()
+      {
+        return supplier.get().hasNext();
       }
-    }
-    catch (Throwable t) {
-      // assume input is opencensus if something went wrong
-    }
-    return OPEN_CENSUS;
+      @Override
+      public InputRow next()
+      {
+        return supplier.get().next();
+      }
+    });
   }
 
   List<InputRow> readAsList()
@@ -281,6 +219,8 @@ public class OpenCensusProtobufReader implements InputEntityReader
   @Override
   public CloseableIterator<InputRowListPlusRawValues> sample() throws IOException
   {
-    return read().map(row -> InputRowListPlusRawValues.of(row, ((MapBasedInputRow) row).getEvent()));
+    try (CloseableIterator<InputRow> iterator = read()) {
+      return iterator.map(row -> InputRowListPlusRawValues.of(row, ((MapBasedInputRow) row).getEvent()));
+    }
   }
 }
